+++ src/programs/obsImpact.f90
@@ -18,6 +18,7 @@ program midas_obsimpact
   !
   ! :Purpose: Main program for Observation Impact computation (FSOI)
   !
+  use version_mod
   use codePrecision_mod
   use ramDisk_mod
   use utilities_mod
@@ -82,12 +83,7 @@ program midas_obsimpact
 
   istamp = exdb('OBSIMPACT','DEBUT','NON')
 
-  write(*,'(/,' //  &
-        '3(" *****************"),/,' //                   &
-        '14x,"-- START OF MIDAS-OBSIMPACT           --",/,' //   &
-        '14x,"-- Calculation of observation impact  --",/, ' //  &
-        '14x,"-- VAR Revision number ",a," --",/,' //  &
-        '3(" *****************"))') 'GIT-REVISION-NUMBER-WILL-BE-ADDED-HERE'
+  call ver_printNameAndVersion('obsImpact','Calculation of observation impact')
 
   ! MPI initilization
   call mpi_initialize